requirements.txt
transformers
streamlit
torch
pandas
ffmpeg-python
yt_dlp
faster_whisper
numpy
datasets